Move PreferenceSerializer to Preference unit
